requirements.txt
# Core ML and Data Science
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
scipy==1.10.1

# Network and Real-time Processing
scapy==2.5.0
pyshark==0.4.5

# Model Serialization
joblib==1.2.0
pickle-mixin==1.0.2

# Visualization
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1
networkx==3.1

# Configuration
pyyaml==6.0
python-json-logger==2.0.7

# Utilities
tqdm==4.65.0
colorama==0.4.6
thread6==1.0.0

# Testing and Development
pytest==7.3.1
black==23.3.0
flake8==6.0.0